RAG/cloud-function/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
PyPDF2==3.*
python-dotenv==1.*
google-auth==2.*
requests==2.*
pytz==2023.*
google-cloud-aiplatform[adk,agent-engines]==1.88.0
google-adk
llama-index
tqdm
google-ai-generativelanguage 